Add extremely basic export functionality
If you click the export button, you'll download a file automatically...
it's svg but you have to rename it for it to work this way (at least
with chrome). Styles for the AA charts had to be moved inside the sites_svg
so that they would export along with the rest.